Rename `testAdjustment` file
